backend/pyproject.toml
[tool.poetry]
name = "minigolfmasters-api"
version = "0.1.0"
description = "Mini Golf Masters backend API"
authors = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
gspread = "^6.1.0"
google-auth = "^2.32.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
